+++ CHANGELOG.md
@@ -7,17 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+## 1.0.0b3 - 2020-08-25
+
 ### Added
 
 - Local time now shown along eve time in timer list
 
-### Changed
-
 ### Fixed
 
 - Form validation no longer forgets values for structure and solar system [#1](https://gitlab.com/ErikKalkoken/aa-structuretimers/-/issues/1)
-
 - Users can delete their own timers
+- Will no longer upgrade to Django 3
 
 ## 1.0.0b2 - 2020-08-19